Die Laufzeitregel gilt jetzt auch im geteilten Code — zwei Fundstellen
Beim eigenen Durchgang durch Sources/Shared gefunden: BrowseViews zeigt
die Laufzeit eines Extras ohne jede Pruefung, SeriesView.nebenzeile
ebenso. runtimeSeconds ist bei einer fehlenden Angabe 0, nicht nil —
dort haette also "0 Min." gestanden. Genau die Klasse, die auf tvOS
wochenlang sichtbar war und die CLAUDE.md als Beispiel fuehrt.

Dazu SucheView auf Anzeigeregeln.suchbegriffTaugt; die Schwelle stand
bisher auf jeder Fassung einzeln getippt.

199 Pakettests gruen, alle drei Apple-Ziele gebaut.

+++ Packages/JellyfinKit/Sources/JellyfinKit/Anzeigeregeln.swift
@@ -22,7 +22,12 @@ public enum Anzeigeregeln {
     ///
     /// Auch `nil` und negative Werte fallen heraus: ein Server, der Unsinn
     /// meldet, soll keine Anzeige erzeugen, die wie eine Angabe aussieht.
-    public static func laufzeitZeigen(sekunden: Int?) -> Bool {
+    ///
+    /// `Double`, weil `Item.runtimeSeconds` einer ist — der Server rechnet
+    /// in Ticks, und daraus wird beim Umrechnen eine Kommazahl. Wer hier
+    /// `Int` schreibt, zwingt jeden Aufrufer zu einer Umwandlung und
+    /// verliert dabei die halbe Sekunde, um die es nie ging.
+    public static func laufzeitZeigen(sekunden: Double?) -> Bool {
         guard let sekunden else { return false }
         return sekunden > 0
     }
